Fix uploader-jqueyr throwing error on zip upload, attempt to skip invisible files contained in macOS zip files

+++ zp-core/zp-extensions/uploader_jQuery/class-uploadhandlerZP.php
@@ -207,6 +207,7 @@ class UploadHandlerZP extends UploadHandler {
 		$albumobj = $imageobj = null;
 		$file = new \stdClass();
 		$file->name = $this->get_file_name($uploaded_file, $name, $size, $type, $error, $index, $content_range);
+		$is_zip_upload = false;
 		
 		// zpcms addition
 		$seoname = seoFriendly($name);
@@ -266,8 +267,12 @@ class UploadHandlerZP extends UploadHandler {
 						}
 						$imageobj->save();
 					} else if (is_zip($targetFile)) {
-						unzip($targetFile, $_zp_uploader_targetpath);
-						unlink($targetFile);
+						$zip_success = unzip($targetFile, $_zp_uploader_targetpath);
+						if($zip_success) {
+							$is_zip_upload = true;
+						} else {
+							unlink($targetFile);
+						}
 					} else {
 						$file->error = $error = $this->get_error_message(UPLOAD_ERR_EXTENSION); // invalid file uploaded
 					}
@@ -290,6 +295,8 @@ class UploadHandlerZP extends UploadHandler {
 					} else {
 						$this->handle_image_file($file_path, $file);
 					}
+				} else if ($is_zip_upload) {
+					unlink($targetFile);
 				}
 			} else {
 				$file->size = $file_size;
